Redesign data model
Measurable and Counter represent something which can be probed.
The same Counter can be probed at different Locations.
Counter and Location are Subject for probing.
Subject together with timestamp and value create a Sample.

The printout is not as pretty as before. This can
be fixed by spolitting subjects by tags then by location.